:art: Forenklet useTabs typer
